+++ pkg/commands/commands.go
@@ -56,9 +56,19 @@ import (
 	"github.com/fastly/cli/pkg/commands/logging/syslog"
 	"github.com/fastly/cli/pkg/commands/logtail"
 	"github.com/fastly/cli/pkg/commands/ngwaf"
+	"github.com/fastly/cli/pkg/commands/ngwaf/countrylist"
+	"github.com/fastly/cli/pkg/commands/ngwaf/iplist"
+	"github.com/fastly/cli/pkg/commands/ngwaf/signallist"
+	"github.com/fastly/cli/pkg/commands/ngwaf/stringlist"
+	"github.com/fastly/cli/pkg/commands/ngwaf/wildcardlist"
 	"github.com/fastly/cli/pkg/commands/ngwaf/workspace"
+	wscountrylist "github.com/fastly/cli/pkg/commands/ngwaf/workspace/countrylist"
+	wsiplist "github.com/fastly/cli/pkg/commands/ngwaf/workspace/iplist"
 	"github.com/fastly/cli/pkg/commands/ngwaf/workspace/redaction"
+	wssignallistlist "github.com/fastly/cli/pkg/commands/ngwaf/workspace/signallist"
+	wsstringlistlist "github.com/fastly/cli/pkg/commands/ngwaf/workspace/stringlist"
 	"github.com/fastly/cli/pkg/commands/ngwaf/workspace/virtualpatch"
+	wswildcardlistlist "github.com/fastly/cli/pkg/commands/ngwaf/workspace/wildcardlist"
 	"github.com/fastly/cli/pkg/commands/objectstorage"
 	"github.com/fastly/cli/pkg/commands/objectstorage/accesskeys"
 	"github.com/fastly/cli/pkg/commands/pop"
@@ -398,19 +408,79 @@ func Define( // nolint:revive // function-length
 	loggingSyslogList := syslog.NewListCommand(loggingSyslogCmdRoot.CmdClause, data)
 	loggingSyslogUpdate := syslog.NewUpdateCommand(loggingSyslogCmdRoot.CmdClause, data)
 	ngwafRoot := ngwaf.NewRootCommand(app, data)
-	ngwafWorkspacesRoot := workspace.NewRootCommand(ngwafRoot.CmdClause, data)
-	ngwafWorkspacesCreate := workspace.NewCreateCommand(ngwafWorkspacesRoot.CmdClause, data)
-	ngwafWorkspacesDelete := workspace.NewDeleteCommand(ngwafWorkspacesRoot.CmdClause, data)
-	ngwafWorkspacesGet := workspace.NewGetCommand(ngwafWorkspacesRoot.CmdClause, data)
-	ngwafWorkspacesList := workspace.NewListCommand(ngwafWorkspacesRoot.CmdClause, data)
-	ngwafWorkspacesUpdate := workspace.NewUpdateCommand(ngwafWorkspacesRoot.CmdClause, data)
-	ngwafRedactionRoot := redaction.NewRootCommand(ngwafWorkspacesRoot.CmdClause, data)
+	ngwafWorkspaceRoot := workspace.NewRootCommand(ngwafRoot.CmdClause, data)
+	ngwafWorkspaceCreate := workspace.NewCreateCommand(ngwafWorkspaceRoot.CmdClause, data)
+	ngwafWorkspaceDelete := workspace.NewDeleteCommand(ngwafWorkspaceRoot.CmdClause, data)
+	ngwafWorkspaceGet := workspace.NewGetCommand(ngwafWorkspaceRoot.CmdClause, data)
+	ngwafWorkspaceList := workspace.NewListCommand(ngwafWorkspaceRoot.CmdClause, data)
+	ngwafWorkspaceUpdate := workspace.NewUpdateCommand(ngwafWorkspaceRoot.CmdClause, data)
+	ngwafRedactionRoot := redaction.NewRootCommand(ngwafWorkspaceRoot.CmdClause, data)
 	ngwafRedactionCreate := redaction.NewCreateCommand(ngwafRedactionRoot.CmdClause, data)
 	ngwafRedactionDelete := redaction.NewDeleteCommand(ngwafRedactionRoot.CmdClause, data)
 	ngwafRedactionList := redaction.NewListCommand(ngwafRedactionRoot.CmdClause, data)
 	ngwafRedactionRetrieve := redaction.NewRetrieveCommand(ngwafRedactionRoot.CmdClause, data)
 	ngwafRedactionUpdate := redaction.NewUpdateCommand(ngwafRedactionRoot.CmdClause, data)
-	ngwafVirtualpatchRoot := virtualpatch.NewRootCommand(ngwafWorkspacesRoot.CmdClause, data)
+	ngwafCountryListRoot := countrylist.NewRootCommand(ngwafRoot.CmdClause, data)
+	ngwafCountryListCreate := countrylist.NewCreateCommand(ngwafCountryListRoot.CmdClause, data)
+	ngwafCountryListDelete := countrylist.NewDeleteCommand(ngwafCountryListRoot.CmdClause, data)
+	ngwafCountryListGet := countrylist.NewGetCommand(ngwafCountryListRoot.CmdClause, data)
+	ngwafCountryListList := countrylist.NewListCommand(ngwafCountryListRoot.CmdClause, data)
+	ngwafCountryListUpdate := countrylist.NewUpdateCommand(ngwafCountryListRoot.CmdClause, data)
+	ngwafIPListRoot := iplist.NewRootCommand(ngwafRoot.CmdClause, data)
+	ngwafIPListCreate := iplist.NewCreateCommand(ngwafIPListRoot.CmdClause, data)
+	ngwafIPListDelete := iplist.NewDeleteCommand(ngwafIPListRoot.CmdClause, data)
+	ngwafIPListGet := iplist.NewGetCommand(ngwafIPListRoot.CmdClause, data)
+	ngwafIPListList := iplist.NewListCommand(ngwafIPListRoot.CmdClause, data)
+	ngwafIPListUpdate := iplist.NewUpdateCommand(ngwafIPListRoot.CmdClause, data)
+	ngwafSignalListRoot := signallist.NewRootCommand(ngwafRoot.CmdClause, data)
+	ngwafSignalListCreate := signallist.NewCreateCommand(ngwafSignalListRoot.CmdClause, data)
+	ngwafSignalListDelete := signallist.NewDeleteCommand(ngwafSignalListRoot.CmdClause, data)
+	ngwafSignalListGet := signallist.NewGetCommand(ngwafSignalListRoot.CmdClause, data)
+	ngwafSignalListList := signallist.NewListCommand(ngwafSignalListRoot.CmdClause, data)
+	ngwafSignalListUpdate := signallist.NewUpdateCommand(ngwafSignalListRoot.CmdClause, data)
+	ngwafStringListRoot := stringlist.NewRootCommand(ngwafRoot.CmdClause, data)
+	ngwafStringListCreate := stringlist.NewCreateCommand(ngwafStringListRoot.CmdClause, data)
+	ngwafStringListDelete := stringlist.NewDeleteCommand(ngwafStringListRoot.CmdClause, data)
+	ngwafStringListGet := stringlist.NewGetCommand(ngwafStringListRoot.CmdClause, data)
+	ngwafStringListList := stringlist.NewListCommand(ngwafStringListRoot.CmdClause, data)
+	ngwafStringListUpdate := stringlist.NewUpdateCommand(ngwafStringListRoot.CmdClause, data)
+	ngwafWildcardListRoot := wildcardlist.NewRootCommand(ngwafRoot.CmdClause, data)
+	ngwafWildcardListCreate := wildcardlist.NewCreateCommand(ngwafWildcardListRoot.CmdClause, data)
+	ngwafWildcardListDelete := wildcardlist.NewDeleteCommand(ngwafWildcardListRoot.CmdClause, data)
+	ngwafWildcardListGet := wildcardlist.NewGetCommand(ngwafWildcardListRoot.CmdClause, data)
+	ngwafWildcardListList := wildcardlist.NewListCommand(ngwafWildcardListRoot.CmdClause, data)
+	ngwafWildcardListUpdate := wildcardlist.NewUpdateCommand(ngwafWildcardListRoot.CmdClause, data)
+	ngwafWorkspaceCountryListRoot := wscountrylist.NewRootCommand(ngwafWorkspaceRoot.CmdClause, data)
+	ngwafWorkspaceCountryListCreate := wscountrylist.NewCreateCommand(ngwafWorkspaceCountryListRoot.CmdClause, data)
+	ngwafWorkspaceCountryListDelete := wscountrylist.NewDeleteCommand(ngwafWorkspaceCountryListRoot.CmdClause, data)
+	ngwafWorkspaceCountryListGet := wscountrylist.NewGetCommand(ngwafWorkspaceCountryListRoot.CmdClause, data)
+	ngwafWorkspaceCountryListList := wscountrylist.NewListCommand(ngwafWorkspaceCountryListRoot.CmdClause, data)
+	ngwafWorkspaceCountryListUpdate := wscountrylist.NewUpdateCommand(ngwafWorkspaceCountryListRoot.CmdClause, data)
+	ngwafWorkspaceIPListRoot := wsiplist.NewRootCommand(ngwafWorkspaceRoot.CmdClause, data)
+	ngwafWorkspaceIPListCreate := wsiplist.NewCreateCommand(ngwafWorkspaceIPListRoot.CmdClause, data)
+	ngwafWorkspaceIPListDelete := wsiplist.NewDeleteCommand(ngwafWorkspaceIPListRoot.CmdClause, data)
+	ngwafWorkspaceIPListGet := wsiplist.NewGetCommand(ngwafWorkspaceIPListRoot.CmdClause, data)
+	ngwafWorkspaceIPListList := wsiplist.NewListCommand(ngwafWorkspaceIPListRoot.CmdClause, data)
+	ngwafWorkspaceIPListUpdate := wsiplist.NewUpdateCommand(ngwafWorkspaceIPListRoot.CmdClause, data)
+	ngwafWorkspaceSignalListRoot := wssignallistlist.NewRootCommand(ngwafWorkspaceRoot.CmdClause, data)
+	ngwafWorkspaceSignalListCreate := wssignallistlist.NewCreateCommand(ngwafWorkspaceSignalListRoot.CmdClause, data)
+	ngwafWorkspaceSignalListDelete := wssignallistlist.NewDeleteCommand(ngwafWorkspaceSignalListRoot.CmdClause, data)
+	ngwafWorkspaceSignalListGet := wssignallistlist.NewGetCommand(ngwafWorkspaceSignalListRoot.CmdClause, data)
+	ngwafWorkspaceSignalListList := wssignallistlist.NewListCommand(ngwafWorkspaceSignalListRoot.CmdClause, data)
+	ngwafWorkspaceSignalListUpdate := wssignallistlist.NewUpdateCommand(ngwafWorkspaceSignalListRoot.CmdClause, data)
+	ngwafWorkspaceStringListRoot := wsstringlistlist.NewRootCommand(ngwafWorkspaceRoot.CmdClause, data)
+	ngwafWorkspaceStringListCreate := wsstringlistlist.NewCreateCommand(ngwafWorkspaceStringListRoot.CmdClause, data)
+	ngwafWorkspaceStringListDelete := wsstringlistlist.NewDeleteCommand(ngwafWorkspaceStringListRoot.CmdClause, data)
+	ngwafWorkspaceStringListGet := wsstringlistlist.NewGetCommand(ngwafWorkspaceStringListRoot.CmdClause, data)
+	ngwafWorkspaceStringListList := wsstringlistlist.NewListCommand(ngwafWorkspaceStringListRoot.CmdClause, data)
+	ngwafWorkspaceStringListUpdate := wsstringlistlist.NewUpdateCommand(ngwafWorkspaceStringListRoot.CmdClause, data)
+	ngwafWorkspaceWildcardListRoot := wildcardlist.NewRootCommand(ngwafWorkspaceRoot.CmdClause, data)
+	ngwafWorkspaceWildcardListCreate := wswildcardlistlist.NewCreateCommand(ngwafWorkspaceWildcardListRoot.CmdClause, data)
+	ngwafWorkspaceWildcardListDelete := wswildcardlistlist.NewDeleteCommand(ngwafWorkspaceWildcardListRoot.CmdClause, data)
+	ngwafWorkspaceWildcardListGet := wswildcardlistlist.NewGetCommand(ngwafWorkspaceWildcardListRoot.CmdClause, data)
+	ngwafWorkspaceWildcardListList := wswildcardlistlist.NewListCommand(ngwafWorkspaceWildcardListRoot.CmdClause, data)
+	ngwafWorkspaceWildcardListUpdate := wswildcardlistlist.NewUpdateCommand(ngwafWorkspaceWildcardListRoot.CmdClause, data)
+	ngwafVirtualpatchRoot := virtualpatch.NewRootCommand(ngwafWorkspaceRoot.CmdClause, data)
 	ngwafVirtualpatchList := virtualpatch.NewListCommand(ngwafVirtualpatchRoot.CmdClause, data)
 	ngwafVirtualpatchUpdate := virtualpatch.NewUpdateCommand(ngwafVirtualpatchRoot.CmdClause, data)
 	ngwafVirtualpatchRetrieve := virtualpatch.NewRetrieveCommand(ngwafVirtualpatchRoot.CmdClause, data)
@@ -840,16 +910,74 @@ func Define( // nolint:revive // function-length
 		ngwafRedactionRetrieve,
 		ngwafRedactionUpdate,
 		ngwafRedactionRoot,
+		ngwafCountryListRoot,
+		ngwafCountryListCreate,
+		ngwafCountryListDelete,
+		ngwafCountryListGet,
+		ngwafCountryListList,
+		ngwafCountryListUpdate,
+		ngwafIPListRoot,
+		ngwafIPListCreate,
+		ngwafIPListDelete,
+		ngwafIPListGet,
+		ngwafIPListList,
+		ngwafIPListUpdate,
+		ngwafSignalListRoot,
+		ngwafSignalListCreate,
+		ngwafSignalListDelete,
+		ngwafSignalListGet,
+		ngwafSignalListList,
+		ngwafSignalListUpdate,
+		ngwafStringListRoot,
+		ngwafStringListCreate,
+		ngwafStringListDelete,
+		ngwafStringListGet,
+		ngwafStringListList,
+		ngwafStringListUpdate,
+		ngwafWildcardListCreate,
+		ngwafWildcardListDelete,
+		ngwafWildcardListGet,
+		ngwafWildcardListList,
+		ngwafWildcardListUpdate,
+		ngwafWorkspaceCountryListRoot,
+		ngwafWorkspaceCountryListCreate,
+		ngwafWorkspaceCountryListDelete,
+		ngwafWorkspaceCountryListGet,
+		ngwafWorkspaceCountryListList,
+		ngwafWorkspaceCountryListUpdate,
+		ngwafWorkspaceIPListRoot,
+		ngwafWorkspaceIPListCreate,
+		ngwafWorkspaceIPListDelete,
+		ngwafWorkspaceIPListGet,
+		ngwafWorkspaceIPListList,
+		ngwafWorkspaceIPListUpdate,
+		ngwafWorkspaceSignalListRoot,
+		ngwafWorkspaceSignalListCreate,
+		ngwafWorkspaceSignalListDelete,
+		ngwafWorkspaceSignalListGet,
+		ngwafWorkspaceSignalListList,
+		ngwafWorkspaceSignalListUpdate,
+		ngwafWorkspaceStringListRoot,
+		ngwafWorkspaceStringListCreate,
+		ngwafWorkspaceStringListDelete,
+		ngwafWorkspaceStringListGet,
+		ngwafWorkspaceStringListList,
+		ngwafWorkspaceStringListUpdate,
+		ngwafWorkspaceWildcardListCreate,
+		ngwafWorkspaceWildcardListDelete,
+		ngwafWorkspaceWildcardListGet,
+		ngwafWorkspaceWildcardListList,
+		ngwafWorkspaceWildcardListUpdate,
 		ngwafVirtualpatchList,
 		ngwafVirtualpatchRetrieve,
 		ngwafVirtualpatchRoot,
 		ngwafVirtualpatchUpdate,
-		ngwafWorkspacesRoot,
-		ngwafWorkspacesCreate,
-		ngwafWorkspacesDelete,
-		ngwafWorkspacesGet,
-		ngwafWorkspacesList,
-		ngwafWorkspacesUpdate,
+		ngwafWorkspaceRoot,
+		ngwafWorkspaceCreate,
+		ngwafWorkspaceDelete,
+		ngwafWorkspaceGet,
+		ngwafWorkspaceList,
+		ngwafWorkspaceUpdate,
 		objectStorageRoot,
 		objectStorageAccesskeysRoot,
 		objectStorageAccesskeysCreate,