+++ pkg/commands/ngwaf/ngwaflist/doc.go
@@ -1,0 +1,2 @@
+// Package list contains generic api calls to inspect and manipulate NGWAF account lists.
+package ngwaflist