+++ pkg/commands/ngwaf/wildcardlist/doc.go
@@ -1,0 +1,2 @@
+// Package wildcardlist contains commands to inspect and manipulate NGWAF account-level wildcard lists.
+package wildcardlist