+++ pkg/commands/ngwaf/workspace/countrylist/doc.go
@@ -1,0 +1,2 @@
+// Package countrylist contains commands to inspect and manipulate NGWAF workspace-level country lists.
+package countrylist